Update rows of body
Update rows of body to 180px auto so the header can be fixed
to the height of 180px and main can grow as it needs to.

The problem with the previous values was that the header grows to
be very big even though it doesn't need to. It grows just because
main needs to grow, and this is obviously bad.